refactor(luascript): move lua errors to own file
