+++ src/CMakeLists.txt
@@ -185,6 +185,9 @@ if (HTTP)
 	target_link_libraries(tfslib PRIVATE http)
 endif ()
 
+add_subdirectory(lua)
+target_link_libraries(tfslib PRIVATE lua)
+
 add_custom_target(format COMMAND /usr/bin/clang-format -style=file -i ${tfs_HDR} ${tfs_SRC} ${tfs_MAIN})
 
 if (BUILD_TESTING)
