+++ src/lua/error.h
@@ -1,0 +1,33 @@
+#pragma once
+
+#include <lua.hpp>
+#include <string>
+
+namespace tfs::lua {
+
+enum ErrorCode_t
+{
+	LUA_ERROR_PLAYER_NOT_FOUND,
+	LUA_ERROR_CREATURE_NOT_FOUND,
+	LUA_ERROR_ITEM_NOT_FOUND,
+	LUA_ERROR_THING_NOT_FOUND,
+	LUA_ERROR_TILE_NOT_FOUND,
+	LUA_ERROR_HOUSE_NOT_FOUND,
+	LUA_ERROR_COMBAT_NOT_FOUND,
+	LUA_ERROR_CONDITION_NOT_FOUND,
+	LUA_ERROR_AREA_NOT_FOUND,
+	LUA_ERROR_CONTAINER_NOT_FOUND,
+	LUA_ERROR_VARIANT_NOT_FOUND,
+	LUA_ERROR_VARIANT_UNKNOWN,
+	LUA_ERROR_SPELL_NOT_FOUND,
+};
+
+std::string getErrorDesc(ErrorCode_t code);
+
+int luaErrorHandler(lua_State* L);
+
+void reportError(std::string_view function, std::string_view error_desc, lua_State* L, bool stack_trace);
+void reportError(lua_State* L, std::string_view error_desc);
+void reportError(std::string_view error_desc);
+
+} // namespace tfs::lua
