+++ src/npc.cpp
@@ -7,6 +7,7 @@
 
 #include "game.h"
 #include "pugicast.h"
+#include "lua/error.h"
 
 extern Game g_game;
 extern LuaEnvironment g_luaEnvironment;
@@ -679,7 +680,7 @@ int NpcScriptInterface::luagetDistanceTo(lua_State* L)
 
 	const auto& npc = env->getNpc();
 	if (!npc) {
-		reportErrorFunc(L, tfs::lua::getErrorDesc(LUA_ERROR_THING_NOT_FOUND));
+		tfs::lua::reportError(L, tfs::lua::getErrorDesc(tfs::lua::LUA_ERROR_THING_NOT_FOUND));
 		lua_pushnil(L);
 		return 1;
 	}
@@ -688,7 +689,7 @@ int NpcScriptInterface::luagetDistanceTo(lua_State* L)
 
 	const auto& thing = env->getThingByUID(uid);
 	if (!thing) {
-		reportErrorFunc(L, tfs::lua::getErrorDesc(LUA_ERROR_THING_NOT_FOUND));
+		tfs::lua::reportError(L, tfs::lua::getErrorDesc(tfs::lua::LUA_ERROR_THING_NOT_FOUND));
 		lua_pushnil(L);
 		return 1;
 	}
@@ -763,7 +764,7 @@ int NpcScriptInterface::luaOpenShopWindow(lua_State* L)
 	}
 
 	if (!lua_istable(L, -1)) {
-		reportErrorFunc(L, "item list is not a table.");
+		tfs::lua::reportError(L, "item list is not a table.");
 		tfs::lua::pushBoolean(L, false);
 		return 1;
 	}
@@ -792,7 +793,7 @@ int NpcScriptInterface::luaOpenShopWindow(lua_State* L)
 
 	const auto& player = tfs::lua::getPlayer(L, -1);
 	if (!player) {
-		reportErrorFunc(L, tfs::lua::getErrorDesc(LUA_ERROR_PLAYER_NOT_FOUND));
+		tfs::lua::reportError(L, tfs::lua::getErrorDesc(tfs::lua::LUA_ERROR_PLAYER_NOT_FOUND));
 		tfs::lua::pushBoolean(L, false);
 		return 1;
 	}
@@ -802,7 +803,7 @@ int NpcScriptInterface::luaOpenShopWindow(lua_State* L)
 
 	const auto& npc = tfs::lua::getScriptEnv()->getNpc();
 	if (!npc) {
-		reportErrorFunc(L, tfs::lua::getErrorDesc(LUA_ERROR_CREATURE_NOT_FOUND));
+		tfs::lua::reportError(L, tfs::lua::getErrorDesc(tfs::lua::LUA_ERROR_CREATURE_NOT_FOUND));
 		tfs::lua::pushBoolean(L, false);
 		return 1;
 	}
@@ -820,14 +821,14 @@ int NpcScriptInterface::luaCloseShopWindow(lua_State* L)
 	// closeShopWindow(cid)
 	const auto& npc = tfs::lua::getScriptEnv()->getNpc();
 	if (!npc) {
-		reportErrorFunc(L, tfs::lua::getErrorDesc(LUA_ERROR_CREATURE_NOT_FOUND));
+		tfs::lua::reportError(L, tfs::lua::getErrorDesc(tfs::lua::LUA_ERROR_CREATURE_NOT_FOUND));
 		tfs::lua::pushBoolean(L, false);
 		return 1;
 	}
 
 	const auto& player = tfs::lua::getPlayer(L, 1);
 	if (!player) {
-		reportErrorFunc(L, tfs::lua::getErrorDesc(LUA_ERROR_PLAYER_NOT_FOUND));
+		tfs::lua::reportError(L, tfs::lua::getErrorDesc(tfs::lua::LUA_ERROR_PLAYER_NOT_FOUND));
 		tfs::lua::pushBoolean(L, false);
 		return 1;
 	}
@@ -860,7 +861,7 @@ int NpcScriptInterface::luaDoSellItem(lua_State* L)
 	// doSellItem(cid, itemid, amount, <optional> subtype, <optional> actionid, <optional: default: 1> canDropOnMap)
 	const auto& player = tfs::lua::getPlayer(L, 1);
 	if (!player) {
-		reportErrorFunc(L, tfs::lua::getErrorDesc(LUA_ERROR_PLAYER_NOT_FOUND));
+		tfs::lua::reportError(L, tfs::lua::getErrorDesc(tfs::lua::LUA_ERROR_PLAYER_NOT_FOUND));
 		tfs::lua::pushBoolean(L, false);
 		return 1;
 	}
@@ -952,21 +953,21 @@ int NpcScriptInterface::luaNpcOpenShopWindow(lua_State* L)
 {
 	// npc:openShopWindow(cid, items, buyCallback, sellCallback)
 	if (!lua_istable(L, 3)) {
-		reportErrorFunc(L, "item list is not a table.");
+		tfs::lua::reportError(L, "item list is not a table.");
 		tfs::lua::pushBoolean(L, false);
 		return 1;
 	}
 
 	const auto& player = tfs::lua::getPlayer(L, 2);
 	if (!player) {
-		reportErrorFunc(L, tfs::lua::getErrorDesc(LUA_ERROR_PLAYER_NOT_FOUND));
+		tfs::lua::reportError(L, tfs::lua::getErrorDesc(tfs::lua::LUA_ERROR_PLAYER_NOT_FOUND));
 		tfs::lua::pushBoolean(L, false);
 		return 1;
 	}
 
 	const auto& npc = tfs::lua::getSharedPtr<Npc>(L, 1);
 	if (!npc) {
-		reportErrorFunc(L, tfs::lua::getErrorDesc(LUA_ERROR_CREATURE_NOT_FOUND));
+		tfs::lua::reportError(L, tfs::lua::getErrorDesc(tfs::lua::LUA_ERROR_CREATURE_NOT_FOUND));
 		tfs::lua::pushBoolean(L, false);
 		return 1;
 	}
@@ -1019,14 +1020,14 @@ int NpcScriptInterface::luaNpcCloseShopWindow(lua_State* L)
 	// npc:closeShopWindow(player)
 	const auto& player = tfs::lua::getPlayer(L, 2);
 	if (!player) {
-		reportErrorFunc(L, tfs::lua::getErrorDesc(LUA_ERROR_PLAYER_NOT_FOUND));
+		tfs::lua::reportError(L, tfs::lua::getErrorDesc(tfs::lua::LUA_ERROR_PLAYER_NOT_FOUND));
 		tfs::lua::pushBoolean(L, false);
 		return 1;
 	}
 
 	const auto& npc = tfs::lua::getSharedPtr<Npc>(L, 1);
 	if (!npc) {
-		reportErrorFunc(L, tfs::lua::getErrorDesc(LUA_ERROR_CREATURE_NOT_FOUND));
+		tfs::lua::reportError(L, tfs::lua::getErrorDesc(tfs::lua::LUA_ERROR_CREATURE_NOT_FOUND));
 		tfs::lua::pushBoolean(L, false);
 		return 1;
 	}
